scripts: exp255 order-vs-family matched comparison — tables only (#255)
Per review: the only metric with rigorous uncertainty is AUPRC (cluster-bootstrap
SE), so the conclusion (order ≈ family on every matched subset, all within ±1 SE)
is drawn from it; the likelihoods are single-run point estimates with NO error
bars, reported as suggestive context only. Drops the plot figures (not warranted
without LL error bars) and moves the analysis from plots/ to scripts/ as a
tables-only repro that emits three GitHub-markdown tables: matched AUPRC, matched
LL func/non-func/gap, and the post-#266 online-vs-offline AUPRC sanity.
